fixed fabsf single-double swap

+++ src/defs.h
@@ -82,7 +82,7 @@ using namespace std;        // means std:: not needed for cout, max, etc
 #ifdef SINGLE
   // single-prec, machine epsilon for rounding
   #define EPSILON (float)6e-08
-  #define FABS(x) fabs(x)
+  #define FABS(x) fabsf(x)
   typedef fftwf_complex FFTW_CPX;           //  single-prec has fftwf_*
   typedef fftwf_plan FFTW_PLAN;
   #define FFTW_INIT fftwf_init_threads
@@ -100,7 +100,7 @@ using namespace std;        // means std:: not needed for cout, max, etc
 #else
   // double-precision, machine epsilon for rounding
   #define EPSILON (double)1.1e-16
-  #define FABS(x) fabsf(x)
+  #define FABS(x) fabs(x)
   typedef fftw_complex FFTW_CPX;           // double-prec has fftw_*
   typedef fftw_plan FFTW_PLAN;
   #define FFTW_INIT fftw_init_threads
